pyproject.toml
[project]
name = "phishing-guard"
version = "2.0.0"
description = "Email phishing detection system with ML-based text analysis"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "httpx>=0.27.0",
    "imapclient>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "pytest>=8.0.0",
    # ── Attachment analysis libraries (Phase 6) ──────────
    "pefile>=2023.2.7",       # PE/EXE static analysis
    "olefile>=0.47",           # Legacy Office OLE analysis
    "PyPDF2>=3.0.0",           # PDF object tree analysis
    "python-magic>=0.4.27",   # MIME type detection (Linux/Docker — requires libmagic1)
    "yara-python>=4.5.0",     # YARA pattern matching (requires libyara-dev)
]

[project.optional-dependencies]
# Windows local development: install python-magic-bin instead of python-magic
# to get the libmagic DLL bundled automatically.
# Run: pip install python-magic-bin
windows-dev = [
    "python-magic-bin>=0.4.14",
]
# Future additions (not yet enabled)
extras = [
    # yara-python is now in main deps — moved when YARA engine was implemented
]
# ML model training
ml = [
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
